?Fixes for the major of issues regarding the 2.0.50 release.

+++ bin/ebuild.sh
@@ -3,12 +3,12 @@
 # Distributed under the terms of the GNU General Public License v2
 # $Header$
 
-SANDBOX_PREDICT="${SANDBOX_PREDICT}:/proc/self/maps:/dev/console"
+SANDBOX_PREDICT="${SANDBOX_PREDICT}:/proc/self/maps:/dev/console:/usr/lib/portage/pym"
 SANDBOX_WRITE="${SANDBOX_WRITE}:/dev/shm:${PORTAGE_TMPDIR}"
 SANDBOX_READ="${SANDBOX_READ}:/dev/shm:${PORTAGE_TMPDIR}"
 
 
-if [ "$*" != "depend" ] && [ "$*" != "clean" ]; then
+if [ "$*" != "depend" ] && [ "$*" != "clean" ] && [ "$*" != "nofetch" ]; then
 	if [ -f "${T}/successful" ]; then
 		rm -f "${T}/successful"
 	fi
